Moved ApproveStatus, RequestData, and AppRequestView to Request.swift Thumbs are no longer buttons. Only one thumb appears to indicate the current ApproveStatus
