+++ Project190/StudentAppRequestView.swift
@@ -7,61 +7,16 @@
 
 import SwiftUI
 
-enum ApproveStatus {
-    case approved
-    case denied
-    case unprocessed
-}
-
-struct RequestData: Hashable, Identifiable {
-    let id = UUID()
-    var appName: String
-    var appDescription: String
-    var approved: ApproveStatus
-    var approvedDuration: Float
-    
-    init(appName:String, approved:ApproveStatus) {
-        self.appName = appName
-        self.approved = approved
-        self.appDescription = "Generic App Description"
-        self.approvedDuration = .infinity
-    }
-    
-    func hash(into hasher: inout Hasher) {
-        hasher.combine(appName)
-    }
-}
-
-//View that displays data for a single request.
-struct AppRequestView: View {
-    var request: RequestData
-    var body: some View {
-        ZStack {
-            Image(systemName:"applelogo")
-                .frame(maxWidth:.infinity, alignment:.leading)
-            
-            Text(request.appName)
-                .frame(maxWidth:.infinity, alignment:.center)
-            HStack {
-                Image(systemName: "hand.thumbsup")
-                    .foregroundColor(.green)
-                Image(systemName: "hand.thumbsdown")
-                    .foregroundColor(.red)
-            }
-            .frame(maxWidth:.infinity, alignment:.trailing)
-        }
-    }
-}
-
 struct StudentAppRequestView: View {
     @Environment(\.dismiss) private var dismiss
     @State private var searchAppName: String = ""
     var adminName = "Admin"
     
     @State var appList:[RequestData] = [
-        RequestData(appName: "App 1", approved: ApproveStatus.unprocessed),
-        RequestData(appName: "App 2", approved: ApproveStatus.approved),
-        RequestData(appName: "App 3", approved: ApproveStatus.denied)
+        RequestData(appName: "Unprocessed Request", approved: ApproveStatus.unprocessed),
+        RequestData(appName: "Approved App", approved: ApproveStatus.approved),
+        RequestData(appName: "Temporarily Approved App", approved: ApproveStatus.approvedTemporary),
+        RequestData(appName: "Denied App", approved: ApproveStatus.denied)
     ]
     
     var body: some View {
